+++ addons/overrides/data/optre/sparrowhawk/config.cpp
@@ -36,12 +36,12 @@ class CfgVehicles
 		crew="3AS_Imperial_TIE_Pilot_F";
 		side=1;
 		faction="34thPC_FactionPhantom_Faction";
-		textures[] =
+		hiddenSelectionsTextures[] =
 		{
-			"overrides\data\optre\sparrowhawk\textures\Body_01_co.paa",
-			"overrides\data\optre\sparrowhawk\textures\Body_02_co.paa",
-			"overrides\data\optre\sparrowhawk\textures\Body_03_co.paa",
-			"overrides\data\optre\sparrowhawk\textures\Body_04_co.paa",
+			"\z\34thpc\addons\overrides\data\optre\sparrowhawk\textures\Body_01_co.paa",
+			"\z\34thpc\addons\overrides\data\optre\sparrowhawk\textures\Body_02_co.paa",
+			"\z\34thpc\addons\overrides\data\optre\sparrowhawk\textures\Body_03_co.paa",
+			"\z\34thpc\addons\overrides\data\optre\sparrowhawk\textures\Body_04_co.paa",
 		};
 	};
 };